responsiveness fixes, leftNavigation change
