+++ apps/client/src/components/ui/layouts/MatJournalLayout.tsx
@@ -3,10 +3,9 @@ import { LeftNavigation } from 'pages/matjournal/common/LeftNavigation';
 import { ProtectedBaseLayout } from 'components/ui/layouts/ProtectedBaseLayout';
 
 export default function MatJournalLayout() {
-
   return (
     <ProtectedBaseLayout>
-      <div className="bg-[url('assets/images/sam-mgrdichian5.webp')] bg-[center 50%] pb-5 min-h-screen relative pt-navHeight pl-leftNavWidth">
+      <div className="bg-[url('assets/images/sam-mgrdichian5.webp')] bg-[center 50%] pb-5 min-h-screen relative pt-navHeight overflow-hidden pl-leftNavWidth">
         <LeftNavigation />
         <Outlet />
       </div>
