+++ apps/client/src/pages/matjournal/createSession/subcomponents/DrillingTimePicker.tsx
@@ -12,7 +12,7 @@ export const DrillingTimePicker = () => {
   const { register, watch, setValue } = useFormContext();
 
   return (
-    <div className="w-3/4 lg:w-full createSessionPickerStyle">
+    <div className="createSessionPickerStyle">
       <Drill className="icon" />
       <h1>Drilling Time</h1>
       <div className="flex gap-x-1 w-full justify-center items-center">
